third_party/WebKit/LayoutTests/editing/selection/move-vertically-with-paddings-borders-expected.txt

This test ensures WebKit takes paddings and borders into account when moving vertically.

test 1
PASS selectWord() is "left2"
PASS selectWord() is "right2"
PASS selectWord() is "left1"
PASS selectWord() is "right1"


test 2
PASS selectWord() is "left2"
PASS selectWord() is "right2"
PASS selectWord() is "left1"
PASS selectWord() is "right1"


test 3
PASS selectWord() is "left2"
PASS selectWord() is "right2"
PASS selectWord() is "left1"
PASS selectWord() is "right1"


test 4
PASS selectWord() is "left2"
PASS selectWord() is "right2"
PASS selectWord() is "left1"
PASS selectWord() is "right1"
PASS selectWord() is "left3"
PASS selectWord() is "right3"
PASS selectWord() is "left2"
PASS selectWord() is "right2"


PASS successfullyParsed is true

TEST COMPLETE
